#N/A: Reunify rules into one and define it for multiple apps
