+++ tests/rules/test_omnienv_no_such_command.py
@@ -1,6 +1,6 @@
 import pytest
 
-from thefuck.rules.pyenv_no_such_command import get_new_command, match
+from thefuck.rules.omnienv_no_such_command import get_new_command, match
 from thefuck.types import Command
 
 
@@ -11,7 +11,7 @@ def output(pyenv_cmd):
 
 @pytest.fixture(autouse=True)
 def Popen(mocker):
-    mock = mocker.patch('thefuck.rules.pyenv_no_such_command.Popen')
+    mock = mocker.patch('thefuck.rules.omnienv_no_such_command.Popen')
     mock.return_value.stdout.readlines.return_value = (
         b'--version\nactivate\ncommands\ncompletions\ndeactivate\nexec_\n'
         b'global\nhelp\nhooks\ninit\ninstall\nlocal\nprefix_\n'
@@ -31,6 +31,11 @@ def Popen(mocker):
 ])
 def test_match(script, pyenv_cmd, output):
     assert match(Command(script, output=output))
+
+
+def test_match_goenv_output_quote():
+    """test goenv's specific output with quotes (')"""
+    assert match(Command('goenv list', output="pyenv: no such command 'list'"))
 
 
 @pytest.mark.parametrize('script, output', [
